Make user optional in MoveUserRecordedDataToRegistrationFacility
