+++ lib/tasks/scripts/move_facility_data.rb
@@ -1,0 +1,81 @@
+class MoveFacilityData
+  attr_reader :user, :source_facility, :destination_facility
+
+  def initialize(source_facility, destination_facility, user: nil)
+    @user = user
+    @source_facility = source_facility
+    @destination_facility = destination_facility
+  end
+
+  def fix_patient_data
+    patients = records_to_move(
+      Patient,
+      user_key: :registration_user,
+      facility_key: :registration_facility
+    )
+    fix_pbi_metadata(patients)
+    fix_data_for_relation(patients, facility_key: :registration_facility)
+  end
+
+  def fix_appointment_data
+    fix_data_for_relation(records_to_move(Appointment))
+  end
+
+  def fix_prescription_drug_data
+    fix_data_for_relation(records_to_move(PrescriptionDrug))
+  end
+
+  def fix_blood_pressure_data
+    blood_pressures = records_to_move(BloodPressure)
+
+    fix_data_for_relation(
+      Encounter
+        .includes(:observations)
+        .where(observations: {observable_id: blood_pressures.pluck(:id)})
+    )
+
+    fix_data_for_relation(blood_pressures)
+  end
+
+  def fix_blood_sugar_data
+    blood_sugars = records_to_move(BloodSugar)
+
+    fix_data_for_relation(
+      Encounter
+        .includes(:observations)
+        .where(observations: {observable_id: blood_sugars.pluck(:id)})
+    )
+    fix_data_for_relation(blood_sugars)
+  end
+
+  private
+
+  def fix_data_for_relation(relation, facility_key: :facility)
+    Rails.logger.info "Moving #{relation.count} #{relation.klass} records, for user: #{user.full_name},"\
+                      "to #{destination_facility.name}"
+    updated_records = relation.update(facility_key => destination_facility)
+    updated_records.count
+  end
+
+  def records_to_move(klass, user_key: :user, facility_key: :facility)
+    records = klass.where(facility_key => source_facility)
+
+    return records.where(user_key => user) if user.present?
+    records
+  end
+
+  def fix_pbi_metadata(patients)
+    patient_business_identifiers = patients.map(&:business_identifiers)
+                                           .flatten
+                                           .select { |pbi|
+                                             pbi.metadata == {"assigning_user_id" => user.id,
+                                                              "assigning_facility_id" => source_facility.id}
+                                           }
+    Rails.logger.info "Fixing metadata for #{patient_business_identifiers.count} PatientBusinessIdentifier records,"\
+                      "for user: #{user.full_name}, changing assigning_facility_id to #{destination_facility.name}"
+    updated_metadata = {"assigning_user_id": user.id, "assigning_facility_id": destination_facility.id}
+    patient_business_identifiers.each do |pbi|
+      pbi.update(metadata: updated_metadata)
+    end
+  end
+end
